feat: :sparkles: Added unitary tests for delete question use case

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from "../repositories/question-repository";
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string;
   questionId: string;
 }
 
@@ -9,12 +10,17 @@ interface DeleteQuestionUseCaseResponse {}
 export class DeleteQuestionUseCase {
   constructor(private questionsRepository: QuestionsRepository) {}
   async execute({
+    authorId,
     questionId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId);
 
     if (!question) {
       throw new Error("Question not found");
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error("Not allowed");
     }
 
     await this.questionsRepository.delete(question);
